server/v2/registrations/src/registrations-service/requirements.txt
Eve==2.2.0
Flask-Cors==5.0.0
requests==2.32.3
hypermea-core==0.9.37

# start: added by add-validation
isodate==0.7.2
# end: added by add-validation

# start: added by add-auth
eve-negotiable-auth==0.9.5
PyJWT==2.10.1
cryptography==44.0.0
# end: added by add-auth
